app:: load scripts outside of renderer class
